backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sounddevice
soundfile
numpy
SpeechRecognition
gTTS
pyaudio
langdetect
SQLAlchemy
psycopg2-binary
